server/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
multipart
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
email-validator
